fix up rename file to BufferImageSource
