+++ src/rendering/init.ts
@@ -2,7 +2,7 @@ import { extensions } from '../extensions/Extensions';
 import { AlphaMask } from './mask/alpha/AlphaMask';
 import { ColorMask } from './mask/color/ColorMask';
 import { StencilMask } from './mask/stencil/StencilMask';
-import { BufferImageSource } from './renderers/shared/texture/sources/BufferSource';
+import { BufferImageSource } from './renderers/shared/texture/sources/BufferImageSource';
 import { CanvasSource } from './renderers/shared/texture/sources/CanvasSource';
 import { ImageSource } from './renderers/shared/texture/sources/ImageSource';
 import { VideoSource } from './renderers/shared/texture/sources/VideoSource';
