+++ src/rendering/renderers/shared/texture/sources/BufferImageSource.ts
